correcao de validacao de dado para edicao, correcao de metodo de init do presenter, centralizacao atraves do start de activity, padronizacao de nome de key dos argumentos, impl de onitemClick para editar duvida,

+++ projeto_android/Vestlist/app/src/main/java/com/alex/vestlist/source/StudentSourceContract.java
@@ -75,6 +75,6 @@ interface StudentSourceContract {
      * Calcula a porcentagem de listas de um professor que estão com status true.
      * @return a porcentagem.
      */
-    public float getListCompletedPercent(Teacher teacher);
+    public double getListCompletedPercent(Teacher teacher);
 
 }
